backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.1.7
colorama==0.4.6
email-validator
fastapi==0.116.1
uvicorn==0.35.0
gunicorn==22.0.0
python-dotenv
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.2
typing-inspection==0.4.1
typing_extensions==4.14.1
cryptography==43.0.3
asgiref==3.8.1
SQLAlchemy==2.0.34
psycopg[binary]==3.2.2
geoalchemy2==0.13.3
Shapely==2.1.2
alembic==1.14.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
httpx==0.28.1
typer==0.12.5
redis==5.0.8
requests==2.32.3
